# api/openapi.yaml
openapi: 3.1.0
info:
  title: ZTE C300 Monitoring API
  description: |
    Read-only SNMP collector for ZTE C300 OLTs (C320-derived mappings remain
    in code but are not a separate compatibility guarantee). Live routes
    return PON inventory, ONU detail, uplink topology, Timescale history,
    status/UNI events and unconfigured-ONU discovery. HTTP POST/DELETE only
    touch the collector Redis cache.

    **Multi-OLT:** when the instance is configured with multiple OLTs (the `OLTS`
    env var), every `/api/v1/board/...` and `/api/v1/history/...` path below is
    also available under `/api/v1/olt/{olt_id}/...`, where `{olt_id}` selects
    the target OLT. The configured default OLT continues to answer the bare
    `/api/v1/board/...` and `/api/v1/history/...` paths shown here. `board_id`
    is the physical GPON slot and `pon_id` is validated against that slot's card
    (GTGO=8 or GTGH=16 PON ports). History routes require TimescaleDB.

    API contract version is `v1` (endpoint prefix `/api/v1`). The binary
    version and contract version are exposed via `X-App-Version` and
    `X-API-Version` response headers on every request.
  version: 0.1.0-dev
  contact:
    name: API Support
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8081
    description: Local development server
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/history/runs:
    get:
      summary: List poller collection runs
      description: Duration, PON success/error counts and ONU rows written.
      operationId: listCollectionRuns
      tags: [History]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        '200':
          description: Collection run list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryRunsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/history/status-events:
    get:
      summary: List ONU operational-state transitions
      description: >
        Derived from stored samples. Existing onu_samples rows are not rewritten.
        Also available at `/api/v1/olt/{olt_id}/history/status-events`.
      operationId: listStatusEvents
      tags: [History]
      parameters:
        - name: serial
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 500 }
      responses:
        '200':
          description: Status event list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryStatusEventsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/history/eth-events:
    get:
      summary: List Ethernet UNI flaps
      description: >
        Per-port admin/link changes derived from stored samples.
        Also available at `/api/v1/olt/{olt_id}/history/eth-events`.
      operationId: listEthEvents
      tags: [History]
      parameters:
        - name: serial
          in: query
          schema: { type: string }
        - name: port
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 500 }
      responses:
        '200':
          description: Ethernet event list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryEthEventsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/history/duplicate-serials:
    get:
      summary: List serials seen at more than one port
      description: >
        Timescale only. `scope=current` (default) is the latest finished run.
        `scope=history` is any serial that has changed board/PON/ONU-ID.
        Also available at `/api/v1/olt/{olt_id}/history/duplicate-serials`.
      operationId: listDuplicateSerials
      tags: [History]
      parameters:
        - name: scope
          in: query
          schema:
            type: string
            enum: [current, history]
            default: current
      responses:
        '200':
          description: Duplicate serial report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryDuplicateSerialsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/history/unauth:
    get:
      summary: List unconfigured ONUs last seen by the OLT
      description: >
        Last discovery snapshot. `unsupported` means the firmware walk is not
        confirmed. Does not delete or rewrite onu_samples.
        Also available at `/api/v1/olt/{olt_id}/history/unauth`.
      operationId: listUnauth
      tags: [History]
      responses:
        '200':
          description: Unconfigured-ONU discovery
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryUnauthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        API key for /api/v1 routes. With a single `API_KEY` the key only gates
        access. With the per-tenant `API_USERS` registry, the key also identifies
        the caller (user_id + role): a user sees only the OLTs whose `user_id`
        matches theirs — requesting another tenant's OLT returns 404 (not 403, to
        avoid enumeration) — while an `admin` role sees all OLTs. A missing or
        unknown key returns 401. Startup requires API_KEY or API_USERS.
...
  schemas:
...
    WebResponse:
      type: object
      description: Standard success response wrapper
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 200
        status:
          type: string
          description: Fixed string "success" for 2xx responses (ISP adapter standard)
          example: success
        data: {}
      required:
        - code
        - status
        - data
    ErrorResponse:
      type: object
      description: |
        Standard error response wrapper aligned to the `isp-adapter-standard`
        wiki page. `error_code` is a machine-readable identifier, `data`
        carries either the error message (string) or a structured object with
        `message` + `details`, and `request_id` echoes the request correlation
        ID so operators can correlate across services.
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 400
        status:
          type: string
          description: HTTP reason phrase (e.g. "Bad Request", "Not Found")
          example: Bad Request
        error_code:
          type: string
          description: Machine-readable error code
          enum:
            - VALIDATION_ERROR
            - NOT_FOUND
            - UNAUTHORIZED
            - SNMP_ERROR
            - REDIS_ERROR
            - CONFIG_ERROR
            - INTERNAL_ERROR
          example: VALIDATION_ERROR
        data:
          description: |
            Either the human-readable error message (string) or an object
            containing `message` + `details` when structured context is
            available. Callers should handle both shapes.
          oneOf:
            - type: string
              example: board_id is not a configured GPON slot
            - type: object
              properties:
                message:
                  type: string
                  example: board_id is not a configured GPON slot
                details:
                  type: object
                  additionalProperties: true
                  example:
                    received: "99"
        request_id:
          type: string
          description: Correlation ID (same value as X-Request-ID response header)
          example: d7dhp4pciod1cuth4i80
      required:
        - code
        - status
        - error_code
        - data
...
    CollectionRun:
      type: object
      properties:
        id: { type: integer }
        device_id: { type: string }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
        status: { type: string }
        pons_ok: { type: integer }
        pons_error: { type: integer }
        onus_sampled: { type: integer }
        duration_ms: { type: integer }
        error: { type: string }
...
    StatusEvent:
      type: object
      properties:
        time: { type: string, format: date-time }
        device_id: { type: string }
        serial_number: { type: string }
        board: { type: integer }
        pon: { type: integer }
        onu_id: { type: integer }
        status: { type: string }
        previous_status: { type: string }
        source: { type: string }
    EthEvent:
      type: object
      properties:
        time: { type: string, format: date-time }
        device_id: { type: string }
        serial_number: { type: string }
        board: { type: integer }
        pon: { type: integer }
        onu_id: { type: integer }
        port: { type: integer }
        link_state: { type: string }
        admin_state: { type: string }
        previous_link: { type: string }
        previous_admin: { type: string }
        speed_mbps: { type: [integer, 'null'] }
        source: { type: string }
    UnauthONU:
      type: object
      properties:
        device_id: { type: string }
        serial_number: { type: string }
        board: { type: integer }
        pon: { type: integer }
        onu_type: { type: string }
        discovery_status: { type: string }
        first_seen_at: { type: string, format: date-time }
        last_seen_at: { type: string, format: date-time }
        gone_at: { type: string, format: date-time }
    UnauthList:
      type: object
      properties:
        status:
          type: string
          enum: [ok, empty, unsupported, unavailable]
        oid: { type: string }
        message: { type: string }
        observed_at: { type: string, format: date-time }
        count: { type: integer }
        onus:
          type: array
          items:
            $ref: '#/components/schemas/UnauthONU'
...
    HistoryRunsResponse:
      allOf:
        - $ref: '#/components/schemas/WebResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/CollectionRun'
    HistoryStatusEventsResponse:
      allOf:
        - $ref: '#/components/schemas/WebResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/StatusEvent'
    HistoryEthEventsResponse:
      allOf:
        - $ref: '#/components/schemas/WebResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/EthEvent'
    HistoryDuplicateSerialsResponse:
      allOf:
        - $ref: '#/components/schemas/WebResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/DuplicateSerialList'
    DuplicatePosition:
      type: object
      properties:
        board: { type: integer }
        pon: { type: integer }
        onu_id: { type: integer }
        name: { type: string }
        status: { type: string }
        last_seen: { type: string, format: date-time }
        samples: { type: integer }
        in_latest_run: { type: boolean }
    DuplicateSerial:
      type: object
      properties:
        device_id: { type: string }
        serial_number: { type: string }
        name: { type: string }
        current_count: { type: integer }
        position_count: { type: integer }
        positions:
          type: array
          items:
            $ref: '#/components/schemas/DuplicatePosition'
    DuplicateSerialList:
      type: object
      properties:
        scope: { type: string, enum: [current, history] }
        run_id: { type: integer, format: int64 }
        from: { type: string, format: date-time }
        to: { type: string, format: date-time }
        count: { type: integer }
        serials:
          type: array
          items:
            $ref: '#/components/schemas/DuplicateSerial'
    HistoryUnauthResponse:
      allOf:
        - $ref: '#/components/schemas/WebResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UnauthList'
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 400
            status: Bad Request
            error_code: VALIDATION_ERROR
            data:
              message: board_id is not a configured GPON slot
              details:
                received: "99"
            request_id: d7dhp4pciod1cuth4i80
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 401
            status: Unauthorized
            error_code: UNAUTHORIZED
            data: invalid or missing API key
            request_id: d7dhp4pciod1cuth4i80
...
    ServiceUnavailable:
      description: History store is not configured or is down
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 503
            status: Service Unavailable
            error_code: SERVICE_UNAVAILABLE
            data: history store is not configured
            request_id: d7dhp4pciod1cuth4i80
...